cleanup classes and repos
