+++ .gitignore
@@ -41,5 +41,5 @@ yarn-error.log*
 next-env.d.ts
 
 # prisma
-prisma/generated
-prisma/migrations
+/prisma/generated
+/prisma/migrations
